Add support for `user/repos` endpoint
This endpoint provides some summary info about all the repos a user has
the ability to see.

This is useful because `get_projects` only shows projects a user is
following, and `get_project` can show a project as existing even if it
isn't currently setup. By checking `has_followers` and filtering by org
name, this lets one more easily get a list of repos which exist in an
org, but which one is not following